Got `lightdb_read_state()`working.  move error struct to errors.rs.
